Rename test files to make them run
The files should match the expected name pattern so that are executed
through maven.
